+++ hydromt/_typing/__init__.py
@@ -1,11 +1,6 @@
 """A module for all of the type definitions used in HydroMT."""
 
 from hydromt._typing.crs import CRS
-from hydromt._typing.error import (
-    NoDataException,
-    NoDataStrategy,
-    exec_nodata_strat,
-)
 from hydromt._typing.fsspec_types import FS
 from hydromt._typing.metadata import SourceMetadata
 from hydromt._typing.model_mode import ModelMode
@@ -50,9 +45,6 @@ __all__ = [
     "TotalBounds",
     "XArrayDict",
     "ModelMode",
-    "NoDataStrategy",
-    "NoDataException",
-    "exec_nodata_strat",
     "Variables",
     "Geom",
     "GpdShapeGeom",