+++ hydromt/data_catalog/drivers/dataset/dataset_driver.py
@@ -1,4 +1,4 @@
-"""Driver for Datasets."""
+"""Abstract driver to read datasets."""
 
 import logging
 from abc import ABC, abstractmethod
@@ -12,8 +12,8 @@ from hydromt._typing import (
     TimeRange,
     Variables,
 )
-from hydromt._typing.error import NoDataStrategy
 from hydromt.data_catalog.drivers.base_driver import BaseDriver
+from hydromt.error import NoDataStrategy
 
 logger = logging.getLogger(__name__)