+++ hydromt/data_catalog/drivers/geodataframe/pyogrio_driver.py
@@ -10,10 +10,10 @@ from pyogrio import read_dataframe, read_info, write_dataframe
 from pyproj import CRS
 
 from hydromt._typing import Bbox, Geom, StrPath
-from hydromt._typing.error import NoDataStrategy, exec_nodata_strat
 from hydromt.data_catalog.drivers.geodataframe.geodataframe_driver import (
     GeoDataFrameDriver,
 )
+from hydromt.error import NoDataStrategy, exec_nodata_strat
 
 logger = logging.getLogger(__name__)