+++ hydromt/data_catalog/drivers/raster/raster_xarray_driver.py
@@ -16,12 +16,12 @@ from hydromt._typing import (
     Variables,
     Zoom,
 )
-from hydromt._typing.error import NoDataStrategy, exec_nodata_strat
 from hydromt._utils.unused_kwargs import _warn_on_unused_kwargs
 from hydromt.data_catalog.drivers.preprocessing import PREPROCESSORS
 from hydromt.data_catalog.drivers.raster.raster_dataset_driver import (
     RasterDatasetDriver,
 )
+from hydromt.error import NoDataStrategy, exec_nodata_strat
 
 logger = logging.getLogger(__name__)