+++ hydromt/data_catalog/sources/dataframe.py
@@ -9,13 +9,13 @@ from pydantic import Field
 from pystac import Catalog as StacCatalog
 
 from hydromt._typing import (
-    NoDataStrategy,
     StrPath,
     TimeRange,
 )
 from hydromt.data_catalog.adapters import DataFrameAdapter
 from hydromt.data_catalog.drivers import DataFrameDriver
 from hydromt.data_catalog.sources import DataSource
+from hydromt.error import NoDataStrategy
 
 logger = logging.getLogger(__name__)