+++ hydromt/data_catalog/sources/geodataframe.py
@@ -18,13 +18,13 @@ from pystac import MediaType
 from hydromt._typing import (
     Bbox,
     Geom,
-    NoDataStrategy,
     StrPath,
     TotalBounds,
 )
 from hydromt.data_catalog.adapters.geodataframe import GeoDataFrameAdapter
 from hydromt.data_catalog.drivers import GeoDataFrameDriver
 from hydromt.data_catalog.sources.data_source import DataSource
+from hydromt.error import NoDataStrategy
 from hydromt.gis.gis_utils import _parse_geom_bbox_buffer
 
 logger = logging.getLogger(__name__)