+++ hydromt/data_catalog/uri_resolvers/raster_tindex_resolver.py
@@ -8,13 +8,12 @@ from typing import List, Optional, Union
 import geopandas as gpd
 
 from hydromt._typing import (
-    NoDataStrategy,
     SourceMetadata,
     TimeRange,
     Zoom,
-    exec_nodata_strat,
 )
 from hydromt.data_catalog.uri_resolvers.uri_resolver import URIResolver
+from hydromt.error import NoDataStrategy, exec_nodata_strat
 
 logger = logging.getLogger(__name__)