+++ tests/cli/test_cli.py
@@ -9,8 +9,8 @@ import pytest
 from click.testing import CliRunner
 
 from hydromt import __version__
-from hydromt._typing import NoDataException
 from hydromt.cli.main import main as hydromt_cli
+from hydromt.error import NoDataException
 from hydromt.model.components.grid import GridComponent
 from tests.conftest import TEST_DATA_DIR